wip continue refactor
