+++ src/lib/builder/trace-builder/build-trace-for-single-port-and-net.ts
