+++ src/lib/builder/trace-builder/pcb-routing/pcb-routing-context.ts
@@ -1,10 +1,13 @@
-import { PCBTraceError } from "@tscircuit/soup"
+import { AnySoupElement, PCBTraceError } from "@tscircuit/soup"
 import { BuildContext } from "lib/types"
+import { PcbObstacle } from "./get-pcb-obstacles"
 
 export type PcbRoutingContext = {
   mutable_pcb_errors: PCBTraceError[]
-  source_trace_id?: string
-  pcb_trace_id?: string
-  pcb_terminal_port_ids?: string[]
+  source_trace_id: string
+  pcb_trace_id: string
+  pcb_terminal_port_ids: string[]
   thickness_mm: number
+  elements: AnySoupElement[]
+  pcb_obstacles: PcbObstacle[]
 } & BuildContext
